+++ ElementX/Sources/Other/MapLibre/MapLibreModels.swift
@@ -38,3 +38,7 @@ enum MapLibreError: Error {
     case failedLocatingUser
     case invalidLocationAuthorization
 }
+
+enum MapTilerGeocodingError: Error {
+    case geocodingFailed
+}
