--- ElementX/Sources/Services/Maps/MapTilerStyleBuilder.swift
@@ -1,46 +1,0 @@
-//
-// Copyright 2023 New Vector Ltd
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-import Foundation
-
-enum MapTilerStyle {
-    case light
-    case dark
-}
-
-class MapTilerStyleBuilder {
-    private let appSettings: AppSettings
-    private let authorization = MapTilerAuthorization()
-    
-    static let shared = MapTilerStyleBuilder()
-    
-    private init() {
-        appSettings = ServiceLocator.shared.settings
-    }
-    
-    func dynamicMapURL(for style: MapTilerStyle) -> URL? {
-        let path: String
-        switch style {
-        case .light:
-            path = appSettings.lightTileMapStyleURL
-        case .dark:
-            path = appSettings.darkTileMapStyleURL
-        }
-        
-        guard let url = URL(string: path) else { return nil }
-        return authorization.authorizateURL(url)
-    }
-}
