+++ ElementX/SupportingFiles/Info.plist
@@ -28,6 +28,8 @@
 	<false/>
 	<key>NSCameraUsageDescription</key>
 	<string>To take pictures or videos and send them as a message $(APP_DISPLAY_NAME) needs access to the camera.</string>
+	<key>NSLocationWhenInUseUsageDescription</key>
+	<string>When you share your location to people, $(APP_DISPLAY_NAME) needs access to show them a map.</string>
 	<key>NSMicrophoneUsageDescription</key>
 	<string>To take videos with audio and send them as a message $(APP_DISPLAY_NAME) needs access to the microphone.</string>
 	<key>NSPhotoLibraryAddUsageDescription</key>
